+++ CHANGELOG.md
@@ -15,6 +15,7 @@ https://github.com/nwnxee/unified/compare/build8193.16...HEAD
 - Events: added Spell Interruption events to SpellEvents
 - Events: added skippable Has Feat event to FeatEvents
 - Events: added Journal Open/Close events to JournalEvents
+- Events: added Detect Mode event to StealthEvents
 - Tweaks: `NWNX_TWEAKS_HIDE_PLAYERS_ON_CHAR_LIST`
 - Tweaks: `NWNX_TWEAKS_FIX_ARMOR_DEX_BONUS_UNDER_ONE`
 - Tweaks: `NWNX_TWEAKS_FIX_ITEM_NULLPTR_IN_CITEMREPOSITORY`